pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "the-ledger"
version = "0.1.0"
description = "Agentic Event Store & Enterprise Audit Infrastructure"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.29.0",
    "pydantic>=2.6.0,<3.0",
    "python-dotenv>=1.0.0,<2.0",
    "fastmcp",
]

[project.optional-dependencies]
agents = [
    "langgraph>=0.2.0,<0.3",
    "langchain>=0.2.0,<0.4",
    "langchain-openai>=0.1.0,<1.0",
    "langchain-google-genai>=1.0.0,<3.0",
    "anthropic>=0.30.0,<0.40",
    "reportlab>=4.2.0,<5.0",
    "openpyxl>=3.1.0,<4.0",
    "faker>=24.0.0,<25.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-timeout>=2.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
